Allow extra field parameters to be serialized in JSON mapping
In some cases, a field definition has extra parameters. For example, the
"nested" field type can have the extra parameter "include_in_parent". We
want to include these extra parameters when serializing the objects to
the JSON object we send to Elasticsearch.

This adds a class with a nested field with a couple field parameters set
and then makes sure that the option appears in the dictionary
serialization of that class's mapping.