Do not use "standard" filter when testing against 7.x
The "standard" filter has been removed in 7.x and causes the integration
tests to fail when run against Elasticsearch 7.x.

Per @HonzaKral's comment on #1288, 7.x clusters can still have indices
created in 6.x, when the "standard" filter was valid. To continue
supporting this until an 8.x release, this commit checks the
Elasticsearch version the tests are being run against before
constructing the analyzer that currently uses the "standard" filter. If
the Elasticsearch version is >= 7, the "standard" filter is not used.