Add param passing notes to documentation on Search.delete, fix broken links
The preferred way of passing parameters to `Search.delete` appears to be calling `Search.params` first,
but this is only ever discussed in GitHub issues like #1115 and #1395.

To help anyone else who has stumbled across this, I added a note about this to the documentation.

I also went ahead and updated the links for the `Search.scan` and `FacetedSearch.params` methods to the most recent versions since they were 404ing.